+++ src/components/editor-block/index.stories.js
@@ -7,7 +7,8 @@ import withReadme from 'storybook-readme/with-readme' //eslint-disable-line
 import { EditorBlock, ExampleInput } from '../..'
 import componentReadme from './README.md'
 
-const block = {
+// Example Config
+const exampleBlock = {
   id: 5,
   data: {
     value: 'Hello World.'
@@ -17,15 +18,40 @@ const block = {
   }
 }
 
+class Wrapper extends React.Component {
+
+  constructor(props) {
+    super(props)
+    this.state = {
+      block: exampleBlock
+    }
+    this.onChange = this.onChange.bind(this)
+  }
+
+  onChange(change) {
+    exampleBlock.data = change.data
+    action('onChange')(change)
+    this.setState({ block: exampleBlock })
+  }
+
+  render() {
+    const { block } = this.state
+    return (
+      <EditorBlock
+        block={block}
+      >
+        <ExampleInput
+          block={block}
+          onChange={this.onChange}
+        />
+      </EditorBlock>
+    )
+  }
+
+}
+
 storiesOf('Editor/Editor-Block', module)
   .addDecorator(withReadme(componentReadme))
   .add('default', () => (
-    <EditorBlock
-      block={block}
-    >
-      <ExampleInput
-        block={block}
-        onChange={action('changed')}
-      />
-    </EditorBlock>
+    <Wrapper />
   ))